+++ src/ray.rs
@@ -7,7 +7,7 @@ pub struct Ray {
 }
 
 impl Ray {
-    pub fn at(&self, t: f64) -> Vec3 {
+    pub fn at(&self, t: f32) -> Vec3 {
         return self.orig + t*self.dir;
     }
